requirements.txt
requests==2.25.1
vk_api==11.9.1
pyaimp==0.2.3
PyYAML==5.4.1
